move ledgerbucketwindow into rpcv1 — its only consumer (#881)
With v2's fee window carrying its own inline ring, the generic
ledgerbucketwindow container has exactly one consumer left: v1's
feewindow. Shared internal/ is for code both trees (or the shared
handlers) genuinely consume, so the package moves into rpcv1. Pure
move; the only edit is v1 feewindow's import path.
